scripts/train.py
#!/usr/bin/env python3
"""
Main Training Script for Drone RL
Trains a PPO policy using PufferLib for hover stabilization
"""

import sys
import os
from pathlib import Path
import yaml
import argparse

# Add src to path
sys.path.insert(0, str(Path(__file__).parent.parent / "src"))

# TODO: Import when implemented
# from rl.drone_env import DroneHoverEnv
# from agents.puffer_trainer import PufferTrainer
# from utils.logger import TrainingLogger


def load_config(config_path: str) -> dict:
    """Load YAML configuration file"""
    with open(config_path, 'r') as f:
        return yaml.safe_load(f)


def main():
    parser = argparse.ArgumentParser(description="Train drone RL policy")
    parser.add_argument(
        "--env-config",
        type=str,
        default="configs/env_config.yaml",
        help="Path to environment config"
    )
    parser.add_argument(
        "--train-config",
        type=str,
        default="configs/training_config.yaml",
        help="Path to training config"
    )
    parser.add_argument(
        "--sim-config",
        type=str,
        default="configs/sim_config.yaml",
        help="Path to simulation config"
    )
    parser.add_argument(
        "--resume",
        type=str,
        default=None,
        help="Path to checkpoint to resume from"
    )
    args = parser.parse_args()

    # Load configurations
    print("Loading configurations...")
    env_config = load_config(args.env_config)
    train_config = load_config(args.train_config)
    sim_config = load_config(args.sim_config)

    print(f"Environment: {env_config['environment']['task_type']}")
    print(f"Total timesteps: {train_config['training']['total_timesteps']:,}")
    print(f"Parallel envs: {train_config['pufferlib']['num_envs']}")

    # TODO: Implement training loop
    # Step 1: Create vectorized environments
    # envs = pufferlib.vector.make(
    #     DroneHoverEnv,
    #     num_envs=train_config['pufferlib']['num_envs'],
    #     env_kwargs={'config': env_config, 'sim_config': sim_config}
    # )

    # Step 2: Initialize trainer
    # trainer = PufferTrainer(envs, train_config)

    # Step 3: Train
    # trainer.train(
    #     total_timesteps=train_config['training']['total_timesteps'],
    #     resume_from=args.resume
    # )

    print("\nTODO: Complete implementation in src/agents/puffer_trainer.py")
    print("TODO: Complete implementation in src/rl/drone_env.py")
    print("\nOnce implemented, training will:")
    print("  1. Create 64 parallel simulation environments")
    print("  2. Train PPO policy for hover stabilization")
    print("  3. Log metrics to TensorBoard")
    print("  4. Save checkpoints every 50k steps")
    print("  5. Evaluate policy every 10k steps")


if __name__ == "__main__":
    main()
